requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
babel==2.17.0
    # via tkcalendar
charset-normalizer==3.4.2
    # via reportlab
colorama==0.4.6
    # via pytest
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
et-xmlfile==2.0.0
    # via openpyxl
exceptiongroup==1.3.0
    # via pytest
fonttools==4.58.4
    # via matplotlib
greenlet==3.2.3
    # via sqlalchemy
iniconfig==2.1.0
    # via pytest
keyboard==0.13.5
    # via -r requirements.in
kiwisolver==1.4.8
    # via matplotlib
matplotlib==3.10.3
    # via -r requirements.in
mysql-connector-python==9.3.0
    # via -r requirements.in
numpy==2.2.6
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   pandas
openpyxl==3.1.5
    # via -r requirements.in
packaging==25.0
    # via
    #   matplotlib
    #   pytest
pandas==2.3.0
    # via -r requirements.in
pillow==11.2.1
    # via
    #   -r requirements.in
    #   matplotlib
    #   pystray
    #   reportlab
pluggy==1.6.0
    # via pytest
pygments==2.19.2
    # via pytest
pyparsing==3.2.3
    # via matplotlib
pystray==0.19.5
    # via -r requirements.in
pytest==8.4.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
reportlab==4.4.2
    # via -r requirements.in
six==1.17.0
    # via
    #   pystray
    #   python-dateutil
sqlalchemy==2.0.41
    # via -r requirements.in
tkcalendar==1.6.1
    # via -r requirements.in
tomli==2.2.1
    # via pytest
typing-extensions==4.14.0
    # via
    #   exceptiongroup
    #   sqlalchemy
tzdata==2025.2
    # via pandas
